DistUpgrade: use org.freedesktop.ScreenSaver to inhibit screen lock
Update the screen lock prevention code in DistUpgradeController to use
the DBus interface org.freedesktop.ScreenSaver. This was suggested as
the preferred alternative to existing quirks, which sometimes fail. If
the org.freedesktop.ScreenSaver service cannot be found, display a
message to the user suggesting that they manually ensure the lock screen
is disabled.

The org.freedesktop.ScreenSaver.Inhibit function is implemented such
that the effect only lasts as long as the caller's dbus connection
remains open [1]. This means that when the dist-upgrade completes, and
the user chooses not to reboot immediately, the screen may lock again.
This is a change in the previous behavior, which was originally added
because of LP: #1780342.

Remove the gnome-screensaver quirk and the KDE-specific code in favor of
this approach. But, leave the xdg-screensaver quirk as it may work in
cases such as xscreensaver, where the org.freedesktop.ScreenSaver
interface is not available.

[1] https://people.freedesktop.org/~hadess/idle-inhibition-spec/ch03.html

LP: #1959458