DistUpgrade: add firefox deb2snap entries for appropriate flavors
Kubuntu, Lubuntu, and Ubuntu MATE are shipping the firefox snap in
jammy. Add logic in DistUpgradeQuirks.py to allow providing a list of
metapackages in a deb2snap entry, instead of just a single entry. Then,
continue as normal if at least one metapackage in the list is installed.

Finally, update deb2snap.json so that snap:firefox is installed on the
correct flavors during dist-upgrade.

LP: #1964443