DistUpgrade: do not show lock screen error if XDG_SESSION_TYPE is tty
The lock screen inhibit code is expected to fail on e.g. server because
the DBus service will not exist. The existing check for empty
XDG_SESSION_TYPE is not sufficient, because this variable is set to
'tty' on server.

Fix the condition so that the error message is not shown if
XDG_SESSION_TYPE is empty, or is set to 'tty'.

LP: #1973785