DistUpgradeQuirks: add quirk to prevent flatpak auto-removal if needed
Some Ubuntu flavors are un-seeding flatpak, which may cause the package
to be marked for auto-removal on upgrade. If a user actively uses
flatpak, we want to prevent this.

Add a quirk to check if there are any flatpak remotes configured, or
apps installed, and then mark the package as manually installed if so.
This will prevent the package, and possibly some related integration
packages, from being auto-removed.