DistUpgradeController: migrate to deb822 sources on upgrade
This adds general support for deb822 to ubuntu-release-upgrader.

Add a check to see if the system is still using any .list source files,
and if so, perform migration to deb822 sources at the end of the
upgrade. The migration itself is pretty straight-forward - just re-write
the sources as deb822, and consolidate where appropriate.

Along with this, write deb822 versions of updateSourcesList() and
rewriteSourcesList(). Rather than trying to adapt the
rewriteSourcesList() function as-is, just start from scratch and do not
carry over old and probably no-longer-needed quirks.

NB that upgrades to mantic should only hit the migration path, and the
re-written updateSourcesList() and rewriteSourcesList(), will not really
be used until upgrades from mantic and newer.