DistUpgradeQuirks: make sure font is restored on reboot
To safely restore the font, it should be done on the next boot. Reason
being, if we restore it at the very end of the upgrade, the font will
still be mangled. But, we need to make sure that once the upgrade exits,
something will still take care of it.

Accomplish this by generating a systemd service that will execute in the
user manager only when ~/.config/upgrade-need-font-restore exists.
After successfully restoring the font, the service will delete
~/.config/upgrade-need-font-restore, hence the service should only run
once.