fix: stop TypeError traceback in DistUpgradeQuirks:_set_generic_font
An edge case bug was discovered recently with the release upgrader.
If the environment being upgraded has the XDG_SESSION_TYPE
variable set to something other than 'tty', and SUDO_UID and
PKEXEC_UID also are not set - i.e. logged in as a root user directly
without invoking `sudo`, casting the presumed UID value to an int
fails because both values in the int cast evaluate to None.

This commit fixes bug `LP: #2038900`.